[libc] Fix wctype_t type and update tests (v3)
- Change wctype_t back to unsigned int
- Update tests to validate API behavior instead of implementation details
- Add missing license headers

+++ libc/include/llvm-libc-types/wctype_t.h
@@ -9,6 +9,6 @@
 #ifndef LLVM_LIBC_TYPES_WCTYPE_T_H
 #define LLVM_LIBC_TYPES_WCTYPE_T_H
 
-typedef unsigned short wctype_t;
+typedef unsigned int wctype_t;
 
 #endif // LLVM_LIBC_TYPES_WCTYPE_T_H
